requirements.txt
Django==5.2
bleach==6.2.0
webencodings==0.5.1
channels
channels-redis
django-crispy-forms
crispy-bootstrap5
Pillow
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-decouple==3.8
whitenoise==6.5.0
redis==4.6.0
celery==5.3.1
requests==2.31.0
stripe==7.8.0
razorpay==1.4.2
reportlab==4.0.7